+++ src/python/interpret/glassbox/ebm/ebm.py
@@ -423,7 +423,7 @@ class BaseCoreEBM(BaseEstimator):
 
     def _fit_main(self, native_ebm, main_attr_sets):
         log.debug("Train main effects")
-        self.current_metric_ = self._cyclic_gradient_boost(
+        self.current_metric_, self.main_episode_idx_ = self._cyclic_gradient_boost(
             native_ebm, main_attr_sets, "Main"
         )
         log.debug("Main Metric: {0}".format(self.current_metric_))
@@ -438,10 +438,12 @@ class BaseCoreEBM(BaseEstimator):
     def staged_fit_interactions(self, X, y, inter_indices=[]):
         check_is_fitted(self, "has_fitted_")
 
-        log.debug("Train interactions")
-
+        self.inter_episode_idx_ = 0
         if len(inter_indices) == 0:
+            log.debug("No interactions to train")
             return self
+
+        log.debug("Training interactions")
 
         # Split data into train/val
         X_train, X_val, y_train, y_val = train_test_split(
@@ -488,7 +490,7 @@ class BaseCoreEBM(BaseEstimator):
             )
         ) as native_ebm:
             log.debug("Train interactions")
-            self.current_metric_ = self._cyclic_gradient_boost(
+            self.current_metric_, self.inter_episode_idx_ = self._cyclic_gradient_boost(
                 native_ebm, inter_attr_sets, "Pair"
             )
             log.debug("Interaction Metric: {0}".format(self.current_metric_))
@@ -513,7 +515,10 @@ class BaseCoreEBM(BaseEstimator):
         min_metric = np.inf
         bp_metric = np.inf
         log.debug("Start boosting {0}".format(name))
+        curr_episode_index = 0
         for data_episode_index in range(self.data_n_episodes):
+            curr_episode_index = data_episode_index
+
             if data_episode_index % 10 == 0:
                 log.debug("Sweep Index for {0}: {1}".format(name, data_episode_index))
                 log.debug("Metric: {0}".format(curr_metric))
@@ -521,7 +526,6 @@ class BaseCoreEBM(BaseEstimator):
             if len(attribute_sets) == 0:
                 log.debug("No sets to boost for {0}".format(name))
 
-            log.debug("Start boosting {0}".format(name))
             for index, attribute_set in enumerate(attribute_sets):
                 curr_metric = native_ebm.training_step(
                     index,
@@ -533,6 +537,7 @@ class BaseCoreEBM(BaseEstimator):
                     validation_weights=0,
                 )
 
+            # NOTE: Out of per-feature boosting on purpose.
             min_metric = min(curr_metric, min_metric)
 
             if no_change_run_length == 0:
@@ -541,12 +546,16 @@ class BaseCoreEBM(BaseEstimator):
                 no_change_run_length = 0
             else:
                 no_change_run_length += 1
-            if no_change_run_length >= self.early_stopping_run_length:
+
+            if (
+                self.early_stopping_run_length >= 0
+                and no_change_run_length >= self.early_stopping_run_length
+            ):
                 log.debug("Early break {0}: {1}".format(name, data_episode_index))
                 break
         log.debug("End boosting {0}".format(name))
 
-        return curr_metric
+        return curr_metric, curr_episode_index
 
 
 class CoreEBMClassifier(BaseCoreEBM, ClassifierMixin):
@@ -826,6 +835,13 @@ class BaseEBM(BaseEstimator):
             self.attribute_set_models_.append(averaged_model)
             self.model_errors_.append(model_errors)
 
+        # Get episode indexes for base estimators.
+        self.main_episode_idxs_ = []
+        self.inter_episode_idxs_ = []
+        for estimator in estimators:
+            self.main_episode_idxs_.append(estimator.main_episode_idx_)
+            self.inter_episode_idxs_.append(estimator.inter_episode_idx_)
+
         # Extract feature names and feature types.
         self.feature_names = []
         self.feature_types = []
@@ -844,6 +860,8 @@ class BaseEBM(BaseEstimator):
             X, self.attribute_sets_, self.attribute_set_models_, []
         )
         self._attrib_set_model_means_ = []
+
+        # TODO: Clean this up before release.
         for set_idx, attribute_set, scores in scores_gen:
             score_mean = np.mean(scores)