+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and the versioning is mostly derived from [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v0.1.2] - 2019-05-17
+### Added
+- EBM can now disable early stopping with run length set to -1.
 ### Fixed
 - Pinning scipy, until upstream dependencies are compatible.
+### Changed
+- Clean-up of EBM logging for training.
 
 ## [v0.1.1] - 2019-05-16
 ### Added